reddit-swarm/requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
openai>=1.12.0
pytest>=7.4.0
